vip/requirements.txt
telethon
python-dotenv
aiohttp
python-telegram-bot
pyTelegramBotAPI
Pillow
requests
aiofiles
